+++ lib/src/square/payment_card.dart
@@ -8,20 +8,20 @@ import 'payment_method.dart';
 
 /// https://developer.squareup.com/reference/sdks/web/payments/objects/Card
 /// https://developer.squareup.com/reference/sdks/web/payments/objects/GiftCard
-class Card implements PaymentMethod {
+class PaymentCard implements PaymentMethod {
   final Future Function(HTMLDivElement element) attach;
   @override
   final Future Function() destroy;
   @override
   final Future<TokenResult> Function() tokenize;
 
-  const Card(
+  const PaymentCard(
       {required this.attach, required this.destroy, required this.tokenize});
 }
 
-extension type JSCard._(JSPaymentMethod _) implements JSPaymentMethod {
+extension type JSPaymentCard._(JSPaymentMethod _) implements JSPaymentMethod {
   external JSPromise attach(HTMLDivElement element);
-  Card get toDart => Card(
+  PaymentCard get toDart => PaymentCard(
       attach: (element) => attach(element).toDart,
       destroy: () => destroy().toDart,
       tokenize: () => tokenize().toDart.then(